Moved CSS-related methods to the widget
