+++ src/base.php
@@ -101,7 +101,6 @@ abstract class PLL_Base {
 		if ( $this->model->has_languages() ) {
 			$this->posts = new PLL_CRUD_Posts( $this );
 			$this->terms = new PLL_CRUD_Terms( $this );
-			Switcher_Assets::init();
 
 			// WordPress options.
 			new PLL_Translate_Option( 'blogname', array(), array( 'context' => 'WordPress' ) );
